# Improve chunk delete performance