### Explain the changes
This PR adds yet another index to `objectparts` relation in the hope of increasing the performance.

I also tried do filtering on the NodeJS side but did not see much performance improvement (and I think the performance might worsen if there are too many objects).

**NOW**
```
nbcore=# EXPLAIN ANALYZE SELECT jsonb_build_object('chunk', data->'chunk') as data FROM objectparts WHERE ((data->>'chunk'='644c59f6549898000d99a03f' and data ? 'chunk') and (data->'deleted' IS NULL OR data->'deleted' = 'null'::jsonb));
                                                                      QUERY PLAN
------------------------------------------------------------------------------------------------------------------------------------------------------
 Index Scan using idx_btree_objectparts_chunk_1_deleted_1 on objectparts  (cost=0.12..8.15 rows=1 width=32) (actual time=0.012..0.013 rows=0 loops=1)
   Index Cond: ((data ->> 'chunk'::text) = '644c59f6549898000d99a03f'::text)
 Planning Time: 0.336 ms
 Execution Time: 0.059 ms
(4 rows)
```

**PREVIOUSLY**
```
nbcore=# EXPLAIN ANALYZE SELECT jsonb_build_object('chunk', data->'chunk') as data FROM objectparts WHERE ((data->>'chunk'='644c59f6549898000d99a03f' and data ? 'chunk') and (data->'deleted' IS NULL OR data->'deleted' = 'null'::jsonb));
                                                                    QUERY PLAN
--------------------------------------------------------------------------------------------------------------------------------------------------
 Index Scan using idx_btree_objectparts_chunk on objectparts  (cost=0.56..4.40 rows=1 width=32) (actual time=0.206..2099.544 rows=696211 loops=1)
   Index Cond: ((data ->> 'chunk'::text) = '644c59f6549898000d99a03f'::text)
   Filter: (((data -> 'deleted'::text) IS NULL) OR ((data -> 'deleted'::text) = 'null'::jsonb))
   Rows Removed by Filter: 52422
 Planning Time: 0.287 ms
 Execution Time: 2150.945 ms
(6 rows)
```

### Issues: Fixed #xxx / Gap #xxx
Please see this issue for more details: https://bugzilla.redhat.com/show_bug.cgi?id=2225008


- [ ] Doc added/updated
- [ ] Tests added