@dannyzaken 2023-11-01
Hi @tangledbytes, sorry for the delay :)

Can you specify which md_stores functions should be affected by this? I see many functions where we query chunks using the `$in` operator. How are these queries perform?