@tangledbytes 2023-11-02
TBH, I don't remember at all why I even created this PR :).

I will get back to this as soon as I can.