@tangledbytes 2024-01-16
@dannyzaken I went back to look at the work I did here and tested it again (against the rebased version). The reason I added this index was to improve the performance of the queries where we search of object parts where `DELETED = NULL` is set. As you noted, these queries are pretty common however PG would use this index only in the queries where there is no other attribute being used and has a smaller index attached to it.

I also compared the write performance against NooBaa with and without changes and did not notice and substantial difference in the performance (it should be noted that I am just testing against a DB with ~ 5K object parts and roughly 20G of data uploaded).

cc: @guymguym