@guymguym 2024-01-29
This case seems to be the same data uploaded many times, causing many object parts to reference the same chunk. When many objects like these are deleted by the user, we are left with many deleted parts referencing the same chunk, and then all the queries that try to filter only non-deleted parts (in read/delete object flows) will suffer from having to filter a lot of those parts (in md_store these are find_parts_by_chunk_ids(), find_parts_unreferenced_chunk_ids(), find_parts_chunks_references(), load_parts_objects_for_chunks()).

Instead of adding a new index for this, we could modify the existing index and add the partial filter, but, we still have one last query that we use without the deleted filter in db_cleaner - has_any_parts_for_chunk().

The question is if we can change the db_cleaner to allow removal of the chunk even if delete parts still exist - which I think is not trivial, but also maintaining an additional index on our largest table is non trivial.

@tangledbytes @dannyzaken @jackyalbo - what do you think about changing how db_cleaner check works to only consider non deleted parts here -

https://github.com/tangledbytes/noobaa-core/blob/a6c16bbbf0ba98859c21427db170a4aae140b139/src/server/bg_services/db_cleaner.js#L71-L73